+++ src/hu_stuff.h
@@ -86,6 +86,7 @@ extern secretmessage_t hud_secret_message; // "A secret is revealed!" message
 
 extern int map_player_coords, map_level_stats, map_level_time; // [FG] level stats and level time widgets
 extern int hud_level_stats, hud_level_time;
+extern int map_power_timers, hud_power_timers; // [Nugget] Powerup timers
 extern int hud_widget_font;
 extern boolean message_centered; // center messages
 extern boolean message_colorized; // colorize player messages