+++ src/m_menu.c
@@ -3367,10 +3367,10 @@ enum {
   stat1_rednum,
   stat1_graypcnt,
   stat1_keys,
-  stat1_stub1,
   stat1_title2,
   stat1_stats,
   stat1_time,
+  stat1_powers, // [Nugget] Powerup timers
   stat1_healthr,
   stat1_healthy,
   stat1_healthg,
@@ -3390,12 +3390,11 @@ setup_menu_t stat_settings1[] =  // Status Bar and HUD Settings screen
   {"GRAY %"            ,S_YESNO|S_COSMETIC,m_null,M_X,M_Y+stat1_graypcnt*M_SPC, {"sts_pct_always_gray"}},
   {"SINGLE KEY DISPLAY",S_YESNO|S_COSMETIC,m_null,M_X,M_Y+stat1_keys*M_SPC,     {"sts_traditional_keys"}},
 
-  {"",S_SKIP,m_null,M_X,M_Y+stat1_stub1*M_SPC},
-
   {"HEADS-UP DISPLAY"  ,S_SKIP|S_TITLE,m_null,M_X,M_Y+stat1_title2*M_SPC},
 
   {"SHOW LEVEL STATS"  ,S_YESNO|S_COSMETIC,m_null,M_X,M_Y+stat1_stats*M_SPC, {"hud_level_stats"}},
   {"SHOW LEVEL TIME"   ,S_YESNO|S_COSMETIC,m_null,M_X,M_Y+stat1_time*M_SPC,  {"hud_level_time"}},
+  {"SHOW POWERUP TIMERS",S_YESNO|S_COSMETIC,m_null,M_X,M_Y+stat1_powers*M_SPC,{"hud_power_timers"}}, // [Nugget]
   {"HEALTH LOW/OK"     ,S_NUM|S_COSMETIC,m_null,M_X,M_Y+stat1_healthr*M_SPC, {"health_red"}},
   {"HEALTH OK/GOOD"    ,S_NUM|S_COSMETIC,m_null,M_X,M_Y+stat1_healthy*M_SPC, {"health_yellow"}},
   {"HEALTH GOOD/EXTRA" ,S_NUM|S_COSMETIC,m_null,M_X,M_Y+stat1_healthg*M_SPC, {"health_green"}},
@@ -3601,6 +3600,7 @@ enum {
   auto1_pointer,
   auto1_stats,
   auto1_time,
+  auto1_powers,
   auto1_stub2,
   auto1_title3,
   auto1_smooth,
@@ -3629,6 +3629,7 @@ setup_menu_t auto_settings1[] =  // 1st AutoMap Settings screen
   {"Coords follow pointer",S_YESNO ,m_null,M_X,M_Y+auto1_pointer*M_SPC, {"map_point_coord"}},  // killough 10/98
   {"Show level stats"     ,S_YESNO ,m_null,M_X,M_Y+auto1_stats*M_SPC,   {"map_level_stats"}},
   {"Show level time"      ,S_YESNO ,m_null,M_X,M_Y+auto1_time*M_SPC,    {"map_level_time"}},
+  {"Show powerup timers"  ,S_YESNO ,m_null,M_X,M_Y+auto1_powers*M_SPC,  {"map_power_timers"}}, // [Nugget]
 
   {"",S_SKIP,m_null,M_X,M_Y+auto1_stub2*M_SPC},